Create Version_01
I put another turret on there and added a loop to use how many turrets there are. But there is a problem of over complicating the signals to tell the ship when it is no longer shooting.
Next Step: make the step Variable adujustable. So that it no longer just creates a circle and sets the points equidistant from each other